#409 Reorganize type alias definitions
