+++ neclib/__init__.py
@@ -69,8 +69,6 @@ from . import parameters  # noqa: F401, E402
 from . import recorders  # noqa: F401, E402
 from . import safety  # noqa: F401, E402
 from . import simulators  # noqa: F401, E402
-from . import typing  # noqa: F401, E402
-from . import units  # noqa: F401, E402
 from . import utils  # noqa: F401, E402
 
 # Aliases
