+++ neclib/recorders/file_writer.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from typing import Optional
 
 from .. import get_logger
-from ..typing import TextLike
+from ..core.type_aliases import TextLike
 from .writer_base import Writer
