+++ neclib/recorders/necstdb_writer.py
@@ -8,7 +8,7 @@ from typing import Any, Callable, Dict, Iterable, List, Optional, Sequence, Tupl
 import necstdb
 
 from .. import get_logger
-from ..typing import TextLike
+from ..core.type_aliases import TextLike
 from .writer_base import Writer
 
 TextData = Union[TextLike, List[TextLike]]
