+++ neclib/utils/math_utils.py
@@ -4,12 +4,12 @@ __all__ = ["clip", "frange", "discretize", "counter", "ConditionChecker"]
 
 import itertools
 import math
-from typing import Generator, Literal, Optional, TypeVar, overload
+from typing import Generator, Literal, Optional, TypeVar, Union, overload
 
+import astropy.units as u
 import numpy as np
 
-from ..typing import QuantityValue, SupportsComparison
-
+from ..core.type_aliases import DimensionLess, SupportsComparison
 
 T = TypeVar("T", bound=SupportsComparison)
 
@@ -63,12 +63,12 @@ def clip(
 
 
 def frange(
-    start: QuantityValue,
-    stop: QuantityValue,
-    step: Optional[QuantityValue] = None,
+    start: Union[u.Quantity, DimensionLess],
+    stop: Union[u.Quantity, DimensionLess],
+    step: Optional[Union[u.Quantity, DimensionLess]] = None,
     *,
     inclusive: bool = False,
-) -> Generator[QuantityValue, None, None]:
+) -> Generator[Union[u.Quantity, DimensionLess], None, None]:
     """Float version of built-in ``range``, with support for including stop value.
 
     Parameters
@@ -97,7 +97,7 @@ def frange(
 
     """
     if step is None:
-        unity: QuantityValue = 1 * getattr(start, "unit", 1)
+        unity: Union[u.Quantity, DimensionLess] = 1 * getattr(start, "unit", 1)
         step = unity
     if inclusive:
         num = -1 * np.ceil((start - stop) / step) + 1
