+++ neclib/utils/quantity_utils.py
@@ -16,7 +16,7 @@ from typing import Any, Dict, Hashable, List, Optional, Union, overload
 
 import astropy.units as u
 
-from ..typing import AngleUnit, Number, UnitType
+from ..core.type_aliases import AngleUnit, DimensionLess, UnitType
 
 
 def angle_conversion_factor(original: AngleUnit, to: AngleUnit) -> float:
@@ -177,8 +177,10 @@ def quantity2builtin(
 
 
 def dAz2dx(
-    az: Union[Number, u.Quantity], el: Union[Number, u.Quantity], unit: UnitType
-) -> Number:
+    az: Union[DimensionLess, u.Quantity],
+    el: Union[DimensionLess, u.Quantity],
+    unit: UnitType,
+) -> DimensionLess:
     az, el = list(
         map(lambda z: z.to_value(unit) if isinstance(z, u.Quantity) else z, [az, el])
     )
@@ -187,8 +189,10 @@ def dAz2dx(
 
 
 def dx2dAz(
-    x: Union[Number, u.Quantity], el: Union[Number, u.Quantity], unit: UnitType
-) -> Number:
+    x: Union[DimensionLess, u.Quantity],
+    el: Union[DimensionLess, u.Quantity],
+    unit: UnitType,
+) -> DimensionLess:
     x, el = list(
         map(lambda z: z.to_value(unit) if isinstance(z, u.Quantity) else z, [x, el])
     )
@@ -228,7 +232,7 @@ def get_quantity(*, default_unit: Optional[UnitType]) -> _GetQuantity:
 
 @overload
 def get_quantity(
-    *value: Union[str, Number, u.Quantity], unit: Optional[UnitType]
+    *value: Union[str, DimensionLess, u.Quantity], unit: Optional[UnitType]
 ) -> u.Quantity:
     ...
